Racheal's version of the game so far
Ok so here are all the files I have for the game thus far. There's a handy makefile and everything is now organized. I commented everything so you can follow along and see where changes need to be made. For some reason changing the vector of bricks to a list created a lot of glitches so if we can't fix it we might consider going back to the vector.